solve render delete button after deleteng transaction

+++ src/pages/Home/Home.js
@@ -70,12 +70,7 @@ Home.prototype.render = async function (parent) {
 
   const transactions = await this.pullAllTransaction()
 
-  const transactionsWithDeleteProperty = transactions.map((transaction) => ({
-    ...transaction,
-    delete: `<button id="${transaction.id}" class="remove-transaction"><svg xmlns="http://www.w3.org/2000/svg" height="16" width="14" viewBox="0 0 448 512"><!--!Font Awesome Free 6.5.1 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license/free Copyright 2023 Fonticons, Inc.--><path d="M135.2 17.7L128 32H32C14.3 32 0 46.3 0 64S14.3 96 32 96H416c17.7 0 32-14.3 32-32s-14.3-32-32-32H320l-7.2-14.3C307.4 6.8 296.3 0 284.2 0H163.8c-12.1 0-23.2 6.8-28.6 17.7zM416 128H32L53.2 467c1.6 25.3 22.6 45 47.9 45H346.9c25.3 0 46.3-19.7 47.9-45L416 128z"/></svg></button>`,
-  }))
-
-  this.transactionsTable = new UniversalTable(transactionsWithDeleteProperty, {
+  this.transactionsTable = new UniversalTable(transactions, {
     headers: [
       {
         name: 'category',
@@ -98,12 +93,17 @@ Home.prototype.render = async function (parent) {
         sort: (a, b) => new Date(a.date).getTime() - new Date(b.date).getTime(),
       },
       { name: 'type', title: 'Type', sortBy: false },
+      { name: 'delete', title: 'Delete' },
     ],
     onClick: async (event, clickedTransaction) => {
-      if (event.target.tagName === 'BUTTON' || event.target.closest(`button.remove-transaction`)) {
+      if (
+        (event.target.tagName === 'BUTTON' && event.target.classList.contains('remove-transaction')) ||
+        event.target.closest(`button.remove-transaction`)
+      ) {
         if (confirm('Are you sure you want to delete this transaction?')) {
           await deleteTransaction(clickedTransaction.id)
-          Router.navigate(ROUTES_NAMES.HOME)
+          const transactions = await this.pullAllTransaction()
+          this.transactionsTable.updateTable(transactions)
         } else return null
       } else return null
     },
@@ -166,6 +166,7 @@ Home.prototype.pullAllTransaction = async function () {
         category: t.category ? (await getDoc(t.category)).data().name : null,
         date: t.date.toDate().toLocaleString(),
         comment: !t.comment ? 'Empty' : t.comment,
+        delete: `<button id="${t.id}" class="remove-transaction"><svg xmlns="http://www.w3.org/2000/svg" height="16" width="14" viewBox="0 0 448 512"><!--!Font Awesome Free 6.5.1 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license/free Copyright 2023 Fonticons, Inc.--><path d="M135.2 17.7L128 32H32C14.3 32 0 46.3 0 64S14.3 96 32 96H416c17.7 0 32-14.3 32-32s-14.3-32-32-32H320l-7.2-14.3C307.4 6.8 296.3 0 284.2 0H163.8c-12.1 0-23.2 6.8-28.6 17.7zM416 128H32L53.2 467c1.6 25.3 22.6 45 47.9 45H346.9c25.3 0 46.3-19.7 47.9-45L416 128z"/></svg></button>`,
       }
     })
   )
